refactor(package) : user 관련 패키지 생성
