+++ src/main/java/roomescape/model/user/UserEmail.java
@@ -1,0 +1,13 @@
+package roomescape.model.user;
+
+public class UserEmail {
+    private String email;
+
+    public UserEmail(String email) {
+        this.email = email;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+}
