+++ src/main/java/roomescape/model/user/UserPassword.java
@@ -1,0 +1,13 @@
+package roomescape.model.user;
+
+public class UserPassword {
+    private String password;
+
+    public UserPassword(String password) {
+        this.password = password;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+}
