+++ src/test/java/roomescape/model/UserNameTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.model.user.UserName;
 
 class UserNameTest {
